+++ Reversi/composeApp/src/jvmMain/kotlin/Main.kt
@@ -1,5 +1,3 @@
-package org.example.reversi.ui
-
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -22,6 +20,7 @@ import reversi.composeapp.generated.resources.Res
 import reversi.composeapp.generated.resources.reversi
 import java.awt.Frame
 
+@Suppress("AssignedValueIsNeverRead")
 fun main() = application {
     val windowState = rememberWindowState(
         size = DpSize(800.dp, 600.dp),
@@ -38,7 +37,7 @@ fun main() = application {
         var currentScreen by remember { mutableStateOf(Screen.START) }
 
         when (currentScreen) {
-            Screen.START -> StartScreen(
+            Screen.START -> _root_ide_package_.reversi_ui.StartScreen(
                 onEnterGame = { currentScreen = Screen.GAME },
                 onCreateGame = { currentScreen = Screen.CREATE },
                 onResolution = { currentScreen = Screen.RESOLUTION }
@@ -46,14 +45,14 @@ fun main() = application {
 
             Screen.CREATE -> {
                 // desenha a StartScreen por baixo (mantendo o background/visual igual ao Join dialog)
-                StartScreen(
+                _root_ide_package_.reversi_ui.StartScreen(
                     onEnterGame = { currentScreen = Screen.GAME },
                     onCreateGame = { currentScreen = Screen.CREATE },
                     onResolution = { currentScreen = Screen.RESOLUTION }
                 )
 
                 // sobrepõe o diálogo de criar jogo
-                CreateGameScreen(
+                _root_ide_package_.reversi_ui.CreateGameScreen(
                     onConfirm = { _, _ ->
                         // navegar para jogo (poderia passar parâmetros ao ViewModel conforme necessário)
                         currentScreen = Screen.GAME
@@ -62,7 +61,7 @@ fun main() = application {
                 )
             }
 
-            Screen.GAME -> GameApp()
+            Screen.GAME -> _root_ide_package_.reversi_ui.GameApp()
 
             Screen.RESOLUTION -> {
                 Box(
@@ -70,7 +69,7 @@ fun main() = application {
                         .fillMaxSize()
                         .background(Color(0xFF0F7A0F))
                 ) {
-                    ResolutionScreen(
+                    _root_ide_package_.reversi_ui.ResolutionScreen(
                         onBack = { currentScreen = Screen.START },
                         onSetWindowSize = { wDp: Dp, hDp: Dp ->
                             windowState.size = DpSize(wDp, hDp)
@@ -79,11 +78,10 @@ fun main() = application {
                         onToggleFullscreen = { fullscreen ->
                             val frame = Frame.getFrames().firstOrNull { it.title == "Reversi" }
                             frame?.let {
-                                if (fullscreen) {
+                                if (fullscreen)
                                     it.extendedState = Frame.MAXIMIZED_BOTH
-                                } else {
+                                else
                                     it.extendedState = Frame.NORMAL
-                                }
                             }
                         }
                     )
@@ -99,3 +97,16 @@ enum class Screen {
     GAME,
     RESOLUTION
 }
+
+// O jogo ta bonito e bem! O front end code ta bom
+// A parte de importar pode tar cook e algumas coisas q n consigo fazer (compreensivel)
+// Cenas da logica do jogo (passar a jogada) (consecutive passes transition to end state)
+
+// If current passes == Max consecutive passes -> transition state
+
+// BASE DE DADOS (MongoDB) atualizacao automatica a partir de la + atualizacao manual
+
+// Haver lobby... lobby ta conectado a base de dados
+// No lobby tem jogos guardados, preview do tabuleiro, proximo jogador, acesso ao jogo?
+// O q acontece no lobby?
+// no need for lobby? (extra)
