+++ Reversi/composeApp/src/jvmMain/kotlin/reversi_ui/GameFileManager.kt
@@ -1,4 +1,4 @@
-package org.example.reversi.ui
+package reversi_ui
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
