+++ Reversi/composeApp/src/jvmMain/kotlin/reversi_ui/JoinGameScreen.kt
@@ -1,4 +1,4 @@
-package org.example.reversi.ui
+package reversi_ui
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.heightIn
