+++ Reversi/composeApp/src/jvmMain/kotlin/reversi_ui/StartScreen.kt
@@ -1,4 +1,4 @@
-package org.example.reversi.ui
+package reversi_ui
 
 import androidx.compose.animation.core.LinearEasing
 import androidx.compose.animation.core.RepeatMode
@@ -7,23 +7,25 @@ import androidx.compose.animation.core.infiniteRepeatable
 import androidx.compose.animation.core.rememberInfiniteTransition
 import androidx.compose.animation.core.tween
 import androidx.compose.foundation.background
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Button
-import androidx.compose.material.ButtonDefaults
 import androidx.compose.material.Card
-import androidx.compose.material.LocalContentAlpha
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.getValue
@@ -31,54 +33,47 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Shadow
 import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.graphics.Shadow
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.input.pointer.PointerEventType
+import androidx.compose.ui.input.pointer.onPointerEvent
 import androidx.compose.ui.layout.onSizeChanged
+import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.sp
+import org.jetbrains.compose.resources.painterResource
+import reversi.composeapp.generated.resources.Res
+import reversi.composeapp.generated.resources.reversi
 
 @Composable
-fun TranslucentButton(
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-    shape: Shape = RoundedCornerShape(12.dp),
-    backgroundAlpha: Float = 0.75f,
-    contentAlpha: Float = 0.6f,
-    content: @Composable RowScope.() -> Unit
+fun StartScreen(
+    onEnterGame: (String) -> Unit,
+    onCreateGame: () -> Unit,
+    onResolution: () -> Unit
 ) {
-    val translucentColors = ButtonDefaults.buttonColors(
-        backgroundColor = Color(0xFF2F2F2F).copy(alpha = backgroundAlpha),
-        contentColor = Color.White
-    )
+    val showDialog = remember { mutableStateOf(false) }
 
-    Button(
-        onClick = onClick,
-        modifier = modifier,
-        shape = shape,
-        colors = translucentColors,
-        content = {
-            CompositionLocalProvider(LocalContentAlpha provides contentAlpha) {
-                content()
-            }
-        }
-    )
-}
-@Composable
-fun ResolutionScreen(
-    onBack: () -> Unit,
-    onSetWindowSize: (widthDp: Dp, heightDp: Dp) -> Unit,
-    onToggleFullscreen: (Boolean) -> Unit // novo callback para fullscreen
-) {
-    // animação de gradiente horizontal (igual à StartScreen)
+    if (showDialog.value) {
+        EnterGameDialog(
+            onConfirm = { name ->
+                showDialog.value = false
+                onEnterGame(name)
+            },
+            onDismiss = { showDialog.value = false }
+        )
+    }
+
+    // animação de gradiente horizontal
     val transition = rememberInfiniteTransition()
     val animProgress by transition.animateFloat(
         initialValue = -1f,
@@ -90,26 +85,24 @@ fun ResolutionScreen(
     )
 
     var boxWidthPx by remember { mutableStateOf(0) }
-    var isFullscreen by remember { mutableStateOf(false) } // estado local do fullscreen
-
-    // cantos dos botões proporcionais aos cantos da Box (Box/Card usa 24.dp - aqui metade = 12.dp)
-    val buttonCorner = 12.dp
 
     Box(
         modifier = Modifier
             .fillMaxSize()
-            .background(Color(0xFF0F260F)), // mesmo fundo escuro/verde
+            .background(Color(0xFF0F260F)), // fundo geral escuro/verde
         contentAlignment = Alignment.Center
     ) {
+        // Card maior verticalmente, largura controlada
         Card(
             modifier = Modifier
                 .clip(RoundedCornerShape(24.dp))
                 .padding(16.dp)
                 .fillMaxWidth(0.86f)
-                .height(420.dp), // maior para cima/baixo igual à StartScreen
+                .height(420.dp), // maior para cima/baixo
             elevation = 8.dp,
             backgroundColor = Color.Transparent
         ) {
+            // recolhe largura para calcular o degradê animado
             Box(
                 modifier = Modifier
                     .onSizeChanged { boxWidthPx = it.width }
@@ -129,28 +122,39 @@ fun ResolutionScreen(
                                 Color(0xFF256E29), Color(0xFF246D29), Color(0xFF236C28), Color(0xFF236B28), Color(0xFF226A27),
                                 Color(0xFF216926), Color(0xFF216826), Color(0xFF206725), Color(0xFF206625), Color(0xFF1F6524),
                                 Color(0xFF1F6424), Color(0xFF1E6323), Color(0xFF1D6222), Color(0xFF1D6122), Color(0xFF1C6021),
-                                Color(0xFF1C5F21), Color(0xFF1B5E20), Color(0xFF1B5D20), Color(0xFF1A5C1F), Color(0xFF195B1E)
+                                Color(0xFF1C5F21), Color(0xFF1B5E20), Color(0xFF1B5D20), Color(0xFF1A5C1F), Color(0xFF195B1E),
                             ),
                             start = Offset(x = if (boxWidthPx == 0) 0f else animProgress * boxWidthPx, y = 0f),
                             end = Offset(x = if (boxWidthPx == 0) boxWidthPx.toFloat() else (animProgress * boxWidthPx + boxWidthPx), y = 0f)
                         ),
-                        shape = RoundedCornerShape(24.dp)
+                        shape = RoundedCornerShape(32.dp)
                     )
                     .padding(28.dp)
             ) {
+                Image(
+                    painter = painterResource(Res.drawable.reversi),
+                    contentDescription = "Mini Tabuleiro Reversi",
+                    modifier = Modifier
+                        .size(100.dp)
+                        .align(Alignment.TopStart)
+                        .offset(x = (-16).dp, y = (-16).dp)
+                        .clip(RoundedCornerShape(8.dp)),
+                    contentScale = ContentScale.Crop
+                )
+
                 Column(
                     modifier = Modifier.fillMaxSize(),
                     horizontalAlignment = Alignment.CenterHorizontally,
                     verticalArrangement = Arrangement.spacedBy(18.dp, alignment = Alignment.CenterVertically)
                 ) {
-                    // título (idêntico ao que já tinha)
+                    // título ornamentado "Reversi" (estilo igual ao de ResolutionScreen)
                     Box(
                         modifier = Modifier
                             .fillMaxWidth()
-                            .padding(bottom = 8.dp),
+                            .padding(bottom = 4.dp),
                         contentAlignment = Alignment.Center
                     ) {
-                        val label = "Escolha uma das resoluções desejadas:"
+                        val label = "Reversi"
                         val outlineColor = Color.Black.copy(alpha = 0.72f)
 
                         val mainStyle = TextStyle(
@@ -161,26 +165,27 @@ fun ResolutionScreen(
                                     Color(0xFF2F7F33)
                                 )
                             ),
-                            fontSize = 24.sp,
-                            fontWeight = FontWeight.Bold,
+                            fontSize = 50.sp,
+                            fontWeight = FontWeight.ExtraBold,
                             fontFamily = FontFamily.SansSerif,
                             letterSpacing = 1.2.sp,
                             shadow = Shadow(
                                 color = Color.Black.copy(alpha = 0.35f),
-                                offset = Offset(2f, 2f),
-                                blurRadius = 6f
+                                offset = Offset(3f, 3f),
+                                blurRadius = 8f
                             )
                         )
 
                         val outlineStyle = TextStyle(
                             color = outlineColor,
-                            fontSize = 24.sp,
-                            fontWeight = FontWeight.Bold,
+                            fontSize = 50.sp,
+                            fontWeight = FontWeight.ExtraBold,
                             fontFamily = FontFamily.SansSerif,
                             letterSpacing = 1.2.sp
                         )
 
-                        Text(label, modifier = Modifier.offset(x = (-1).dp, y = (-1).dp), style = outlineStyle, textAlign = TextAlign.Center)
+                        // múltiplas camadas para criar contorno
+                        Text(label, modifier = Modifier.padding(bottom = 0.dp).then(Modifier.offset(x = (-1).dp, y = (-1).dp)), style = outlineStyle, textAlign = TextAlign.Center)
                         Text(label, modifier = Modifier.offset(x = (-1).dp, y = 0.dp), style = outlineStyle, textAlign = TextAlign.Center)
                         Text(label, modifier = Modifier.offset(x = (-1).dp, y = 1.dp), style = outlineStyle, textAlign = TextAlign.Center)
                         Text(label, modifier = Modifier.offset(x = 0.dp, y = (-1).dp), style = outlineStyle, textAlign = TextAlign.Center)
@@ -193,56 +198,116 @@ fun ResolutionScreen(
                             label,
                             style = mainStyle,
                             textAlign = TextAlign.Center,
-                            maxLines = 2,
+                            maxLines = 1,
                             overflow = TextOverflow.Ellipsis
                         )
                     }
 
-                    TranslucentButton(
-                        onClick = { onSetWindowSize(800.dp, 600.dp) },
-                        shape = RoundedCornerShape(buttonCorner)
-                    ) {
-                        Text("800 x 600")
-                    }
+                    AccentButton(
+                        onClick = { showDialog.value = true },
+                        text = "Entrar num jogo",
+                        modifier = Modifier.fillMaxWidth(0.3f)
+                    )
 
-                    TranslucentButton(
-                        onClick = { onSetWindowSize(1024.dp, 768.dp) },
-                        shape = RoundedCornerShape(buttonCorner)
-                    ) {
-                        Text("1024 x 768")
-                    }
+                    AccentButton(
+                        onClick = onCreateGame,
+                        text = "Criar jogo",
+                        modifier = Modifier.fillMaxWidth(0.3f)
+                    )
 
-                    TranslucentButton(
-                        onClick = { onSetWindowSize(1280.dp, 720.dp) },
-                        shape = RoundedCornerShape(buttonCorner)
-                    ) {
-                        Text("1280 x 720")
-                    }
-
-                    TranslucentButton(
-                        onClick = {
-                            val next = !isFullscreen
-                            isFullscreen = next
-                            onToggleFullscreen(next)
-                        },
-                        shape = RoundedCornerShape(buttonCorner)
-                    ) {
-                        Text(if (isFullscreen) "Sair de Tela Cheia" else "Tela Cheia")
-                    }
-
-                    // botão Voltar com cor cinzenta-escura e cantos arredondados proporcionais
-                    Button(
-                        onClick = onBack,
-                        colors = ButtonDefaults.buttonColors(
-                            backgroundColor = Color(0xFF2F2F2F),
-                            contentColor = Color.White
-                        ),
-                        shape = RoundedCornerShape(buttonCorner)
-                    ) {
-                        Text("Voltar")
-                    }
+                    AccentButton(
+                        onClick = onResolution,
+                        text = "Resolução",
+                        modifier = Modifier.fillMaxWidth(0.2f)
+                    )
                 }
             }
         }
+    }
+}
+
+@OptIn(ExperimentalComposeUiApi::class)
+@Composable
+fun AccentButton(
+    text: String,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+    background: Color = Color.White.copy(alpha = 0.06f),
+    borderColor: Color = Color.Black.copy(alpha = 0.75f),
+    shape: Shape = RoundedCornerShape(32.dp),
+    contentPadding: PaddingValues = PaddingValues(horizontal = 14.dp, vertical = 8.dp)
+) {
+    var hovered by remember { mutableStateOf(false) }
+
+    val bg = if (hovered) background.copy(alpha = 0.12f) else background
+    val elevation = if (hovered) 6.dp else 0.dp
+
+    @Suppress("AssignedValueIsNeverRead")
+    Box(
+        modifier = modifier
+            .onPointerEvent(PointerEventType.Enter) {
+                hovered = true
+            }
+            .onPointerEvent(PointerEventType.Exit) {
+                hovered = false
+            }
+            .shadow(elevation, shape)
+            .clip(shape)
+            .background(bg, shape)
+            .border(BorderStroke(2.dp, borderColor), shape)
+            .clickable(
+                onClick = onClick,
+                indication = null,
+                interactionSource = remember { MutableInteractionSource() }
+            )
+            .padding(contentPadding),
+        contentAlignment = Alignment.Center
+    ) {
+        val outlineColor = Color.Black.copy(alpha = 0.72f)
+
+        val mainStyle = TextStyle(
+            brush = Brush.horizontalGradient(
+                colors = listOf(
+                    Color(0xFF9BE49A),
+                    Color(0xFF49A64C),
+                    Color(0xFF2F7F33)
+                )
+            ),
+            fontSize = 16.sp,
+            fontWeight = FontWeight.Medium,
+            fontFamily = FontFamily.SansSerif,
+            letterSpacing = 0.4.sp,
+            shadow = Shadow(
+                color = Color.Black.copy(alpha = 0.25f),
+                offset = Offset(1f, 1f),
+                blurRadius = 6f
+            )
+        )
+
+        val outlineStyle = TextStyle(
+            color = outlineColor,
+            fontSize = 16.sp,
+            fontWeight = FontWeight.Medium,
+            fontFamily = FontFamily.SansSerif,
+            letterSpacing = 0.4.sp
+        )
+
+        // camadas de contorno
+        Text(text, modifier = Modifier.offset(x = (-1).dp, y = (-1).dp), style = outlineStyle)
+        Text(text, modifier = Modifier.offset(x = (-1).dp, y = 0.dp), style = outlineStyle)
+        Text(text, modifier = Modifier.offset(x = (-1).dp, y = 1.dp), style = outlineStyle)
+        Text(text, modifier = Modifier.offset(x = 0.dp, y = (-1).dp), style = outlineStyle)
+        Text(text, modifier = Modifier.offset(x = 0.dp, y = 1.dp), style = outlineStyle)
+        Text(text, modifier = Modifier.offset(x = 1.dp, y = (-1).dp), style = outlineStyle)
+        Text(text, modifier = Modifier.offset(x = 1.dp, y = 0.dp), style = outlineStyle)
+        Text(text, modifier = Modifier.offset(x = 1.dp, y = 1.dp), style = outlineStyle)
+
+        // texto principal com degrade
+        Text(
+            text,
+            style = mainStyle,
+            maxLines = 1,
+            overflow = TextOverflow.Ellipsis
+        )
     }
 }
