+++ Reversi/reversiCore/src/main/kotlin/core/ReversiLogic.kt
@@ -192,74 +192,3 @@ fun Map<Cell, ReversiPiece>.getScore(): ReversiScore {
     return ReversiScore(white, black)
 }
 
-/**
- * Aplica a ação e devolve o novo estado + lista das posições viradas (ordem perto -> longe).
- */
-fun ReversiState.reversiApplyActionWithFlips(action: ReversiAction, board: ReversiBoard): Pair<ReversiState, List<Cell>> {
-    val currentColor = currentTurn
-    val opponentColor = currentColor.invertColor()
-
-    // Pass (sem posição)
-    if (action.position == null) {
-        val nextState = copy(
-            currentTurn = opponentColor,
-            consecutivePasses = consecutivePasses + 1
-        )
-        return Pair(nextState, emptyList())
-    }
-
-    val newPieces = pieces.toMutableMap()
-    val pos = action.position
-
-    // Coloca a peça nova
-    newPieces[pos] = ReversiPiece(pos, currentColor)
-
-    val allFlipped = mutableListOf<Cell>()
-
-    for (dir in ReversiDirection.entries) {
-        var row = pos.row
-        var col = pos.col
-        val toFlip = mutableListOf<Cell>()
-
-        while (true) {
-            row += dir.dx
-            col += dir.dy
-            val cell = Cell(row, col)
-
-            if (!board.isValid(cell)) { // fora do tabuleiro
-                toFlip.clear()
-                break
-            }
-
-            val currentPiece = newPieces[cell] ?: run {
-                toFlip.clear() // encontrou vazio >> não vira nesta direção
-                break
-            }
-
-            if (currentPiece.color == opponentColor) {
-                toFlip += cell
-                continue
-            }
-
-            if (currentPiece.color == currentColor) {
-                // confirma virar >> ficaram em toFlip (já estão em ordem próximo >> longe)
-                allFlipped += toFlip
-            }
-            break
-        }
-    }
-
-    // Aplica flips
-    for (flipPos in allFlipped) {
-        newPieces[flipPos] = ReversiPiece(flipPos, currentColor)
-    }
-
-    val newState = copy(
-        pieces = newPieces,
-        currentTurn = opponentColor,
-        consecutivePasses = 0
-    )
-
-    return Pair(newState, allFlipped)
-}
-
